ch06-chunking/extractors/requirements.txt
pdfplumber>=0.11,<0.12
beautifulsoup4>=4.12,<5.0
readability-lxml>=0.8,<1.0
lxml>=5.0,<6.0
